Spotifyproxy: Audio: Reworked frame purging mechanism.
Instead of making a large purge on the buffer make a few small ones when
it gets full, shrinking it to the half of it's capacity.

This should mitigate the errors with simmultaneous clients that are
behind of ther ones.

This closes mazkolain/spotimc#99 (or at least it should mitigate the
issue largely).